fix: removed old caching logic & refracted imports
